Rename `cluster_points` to `point_reducer`
To keep the file nameing convention consistant.
