+++ panoptes_aggregation/reducers/point_reducer.py
